% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_from}
\alias{dist_from}
\title{This function compute the variation of information
from a given clustering c0 and all the other partition in the space.
Generate space of all partitions using code form fxt library.
The function returns the distances and block dimensions for each partitions generated.}
\usage{
dist_from(clust, return_partitions = FALSE)
}
\arguments{
\item{clust}{clustering of interest}

\item{return_partitions}{boolean, if TRUE the function returns all the partitions in the space}
}
\description{
This function compute the variation of information
from a given clustering c0 and all the other partition in the space.
Generate space of all partitions using code form fxt library.
The function returns the distances and block dimensions for each partitions generated.
}
