% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsLinearModel}
\alias{gibbsLinearModel}
\title{Bayesian linear regression.}
\usage{
gibbsLinearModel(Y0, X0, a_prior, tau_prior, b_prior, Q_prior, n_iter0)
}
\arguments{
\item{tau_prior}{inverse of variance of the normal prior on intercept}

\item{b_prior}{mean vector of the normal multivariate prior for the coefficients}

\item{Q_prior}{covariance matrix of the normal multivariate prior for the coefficients}

\item{n_iter0}{total number of iterations}

\item{Y}{n x 1 vector of response variable}

\item{X}{n x p matrix of covariates (including intercept)}
}
\description{
This function does a gibbs sampling for a linear regression model.
}
