% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vi_distC}
\alias{vi_distC}
\title{This function compute the variation of information
between two clustering}
\usage{
vi_distC(cl1, cl2)
}
\arguments{
\item{cl1}{vector of clustering labels}

\item{cl2}{vector of clustering labels}
}
\description{
This function compute the variation of information
between two clustering
}
